pyproject.toml
[project]
name = "antigravity-telegram"
version = "0.1.0"
description = "Telegram-based Agentic Bridge for Antigravity Mobile Command"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Antigravity Mobile Command" }
]

dependencies = [
    "python-telegram-bot[ext]>=21.0",
    "mcp>=1.0.0",
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
antigravity-telegram = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
